+++ yggdrasil-web/src/api/mod.rs
@@ -2,4 +2,5 @@ pub mod admin;
 pub mod me;
 pub mod scores;
 pub mod universes;
+pub mod user_profiles;
 pub mod users;